requirements.txt
ultralytics>=8.0.0
pillow>=9.0.0
numpy>=1.21.0
easyocr>=1.7.0
torch>=1.9.0
torchvision>=0.10.0 